se paso a React las gráficas, el check de las tareas y mostrar su informacion
