+++ src/insights.css
@@ -1,0 +1,81 @@
+body {
+    font-family: Arial, sans-serif;
+    background-color: #f4f4f4;
+    padding: 20px;
+  }
+  
+  /* Estilos para el botón de regresar */
+  .back-button {
+    background-color: #3498db;
+    border: none;
+    padding: 10px 20px;
+    border-radius: 5px;
+    cursor: pointer;
+    transition: background-color 0.3s ease;
+    margin-bottom: 20px;
+  }
+  
+  .back-button a {
+    text-decoration: none;
+    color: white;
+    font-size: 16px;
+  }
+  
+  .back-button:hover {
+    background-color: #2980b9;
+  }
+  
+  /* Estilos para el contenedor */
+  .container-chart-choose {
+    background-color: white;
+    padding: 20px;
+    border-radius: 10px;
+    box-shadow: 0 4px 8px rgba(0, 0, 0, 0.1);
+    max-width: 400px;
+    margin: auto;
+  }
+  
+  /* Estilos para el formulario */
+  #form-chart {
+    display: flex;
+    flex-direction: column;
+    gap: 15px;
+  }
+  
+  /* Estilos para el select */
+  #select-chart {
+    padding: 10px;
+    border-radius: 5px;
+    border: 1px solid #ccc;
+    font-size: 16px;
+    background-color: #f9f9f9;
+  }
+  
+  /* Estilos para el botón de generar */
+  .generate-button {
+    background-color: #2ecc71;
+    color: white;
+    padding: 10px;
+    border: none;
+    border-radius: 5px;
+    cursor: pointer;
+    font-size: 16px;
+    transition: background-color 0.3s ease;
+  }
+  
+  .generate-button:hover {
+    background-color: #27ae60;
+  }
+  
+  /* Media query para dispositivos móviles */
+  @media (max-width: 600px) {
+    .container-chart-choose {
+      max-width: 100%;
+      padding: 15px;
+    }
+  
+    .back-button {
+      padding: 8px 15px;
+      font-size: 14px;
+    }
+  }
